requirements.txt
django>=2.0.0,<3.0.0
djangorestframework==3.11.0
django-model-utils==4.0.0
django-filter==2.2.0
djangorestframework-api-key==1.4.1
psycopg2==2.8.4
dynaconf[toml]==2.2.3
django-autocomplete-light==3.5.1
pillow==7.0.0
django-cleanup==4.0.0
django-imagekit==4.0.2
django-silk==4.0.1
